replacing cp with cat in some instances to overcome overlayFS bug
"overlayfs: ERROR - failed to whiteout 'majestic.yaml'"

I think this bug affect overlayFS upt o version v20 but this is not tested.

Background: when a file is overwritten with cp on the lower FS the kernel first tryies to white it out (mark as deleted) which fails.
With cat it seems working as it is threated as edit. When a file is edited it first copied to the upper file system then modified.